pyproject.toml
[project]
name = "gatekeeper-core"
version = "0.1.0"
description = "Trading discipline platform — define rules, enforce checklists, journal trades, measure consistency"
requires-python = ">=3.12"
license = "AGPL-3.0-or-later"
dependencies = [
    "fastapi>=0.115.0",
    "uvicorn[standard]>=0.32.0",
    "sqlalchemy[asyncio]>=2.0.36",
    "asyncpg>=0.30.0",
    "alembic>=1.14.0",
    "pydantic-settings>=2.6.0",
    "jinja2>=3.1.4",
    "python-multipart>=0.0.12",
    "itsdangerous>=2.2.0",
    "httpx>=0.28.0",
    "anthropic>=0.39.0",
    "openai>=1.0",
    "slowapi>=0.1.9",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.3.0",
    "pytest-asyncio>=0.24.0",
    "httpx>=0.28.0",
    "coverage>=7.0",
]

[tool.hatch.build.targets.wheel]
packages = ["app"]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[dependency-groups]
dev = [
    "pytest-cov>=7.0.0",
]

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]
